pyproject.toml
[tool.poetry]
name = "uunit_schedule_bot"
version = "0.1.0"
description = ""
authors = [
    "Kamil",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
telebot = "^0.0.4"
peewee = "^3.15.4"
scheduler = "^0.8.2"
pytelegrambotapi = "^4.8.0"
isoweek = "^1.3.3"
python-telegram-bot-pagination = "^0.0.2"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.2"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
